requirements.txt
git+https://github.com/pegasus-isi/pegasus.git@7d70c4a18f4df30090c830c5b07f388df88b44e2#egg=pegasus-wms&subdirectory=lib/pegasus/python
git+https://github.com/isi-vista/saga-tools.git@master#egg=saga-tools
# core requirements
attrs==19.2.0
vistautils==0.20.0
immutablecollections==0.10.0
networkx==2.3
importlib-resources==1.4.0

# extras for development
black==19.3b0
pytest==5.2.0
pytest-custom_exit_code
coverage==4.5.4
pytest-cov==2.8.0
pytest-benchmark==3.2.2
pylint==2.3.1
flake8==3.7.8
mypy==0.701
isort==4.3.21
towncrier
sphinx
sphinx-autodoc-typehints
sphinx_rtd_theme
